Switch to real-valued event times

+++ MockDB/src/edu/depauw/csc480/mockdb/sim/AbstractEvent.java
@@ -1,14 +1,14 @@
 package edu.depauw.csc480.mockdb.sim;
 
 public abstract class AbstractEvent implements Event {
-	private int time;
+	private double time;
 
-	public AbstractEvent(int time) {
+	public AbstractEvent(double time) {
 		this.time = time;
 	}
 
 	@Override
-	public int getTime() {
+	public double getTime() {
 		return time;
 	}
 }
